# shared/contracts/review.openapi.yaml
openapi: "3.1.0"
info:
  title: CIS Review Queue Service
  version: "1.0.0"
  description: Contact Integrity System - Moderation Review Queue API

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Service health check
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /cases:
    get:
      operationId: listCases
      summary: List moderation cases with filters
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_review, resolved, appealed, overturned]
        - name: assigned_to
          in: query
          schema:
            type: string
        - name: user_id
          in: query
          schema:
            type: string
        - name: priority_min
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseListResponse"

    post:
      operationId: createCase
      summary: Create a new moderation case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCaseRequest"
      responses:
        "201":
          description: Case created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{case_id}:
    get:
      operationId: getCase
      summary: Get case details
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Case details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseDetail"
        "404":
          description: Case not found

    patch:
      operationId: updateCase
      summary: Update case (assign, change status)
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCaseRequest"
      responses:
        "200":
          description: Updated case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /cases/{case_id}/actions:
    post:
      operationId: createModerationAction
      summary: Record a moderation action on a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateActionRequest"
      responses:
        "201":
          description: Action recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationAction"

  /cases/{case_id}/appeal:
    post:
      operationId: fileAppeal
      summary: File an appeal for a case
      parameters:
        - name: case_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppealRequest"
      responses:
        "200":
          description: Appeal filed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"

  /stats:
    get:
      operationId: getStats
      summary: Get review queue statistics
      parameters:
        - name: period_days
          in: query
          schema:
            type: integer
            default: 7
      responses:
        "200":
          description: Queue statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueStats"

components:
  schemas:
    HealthResponse:
      type: object
      required: [status, service, version]
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        service:
          type: string
        version:
          type: string

    Case:
      type: object
      required: [id, detection_id, user_id, thread_id, status, priority, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        detection_id:
          type: string
          format: uuid
        user_id:
          type: string
        thread_id:
          type: string
        status:
          type: string
          enum: [open, in_review, resolved, appealed, overturned]
        priority:
          type: integer
        assigned_to:
          type: string
          nullable: true
        resolution:
          type: string
          nullable: true
        resolution_note:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    CaseDetail:
      allOf:
        - $ref: "#/components/schemas/Case"
        - type: object
          properties:
            detection:
              type: object
              description: Associated detection result
            actions:
              type: array
              items:
                $ref: "#/components/schemas/ModerationAction"
            strikes:
              type: array
              items:
                type: object

    CaseListResponse:
      type: object
      required: [cases, total, limit, offset]
      properties:
        cases:
          type: array
          items:
            $ref: "#/components/schemas/Case"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer

    CreateCaseRequest:
      type: object
      required: [detection_id, user_id, thread_id]
      properties:
        detection_id:
          type: string
          format: uuid
        user_id:
          type: string
        thread_id:
          type: string
        priority:
          type: integer
          default: 0

    UpdateCaseRequest:
      type: object
      properties:
        status:
          type: string
          enum: [open, in_review, resolved, appealed, overturned]
        assigned_to:
          type: string
          nullable: true
        resolution:
          type: string
          enum: [confirmed, false_positive, escalated]
        resolution_note:
          type: string

    CreateActionRequest:
      type: object
      required: [actor_id, actor_role, action_type, target_user_id, target_scope, reason_code]
      properties:
        actor_id:
          type: string
        actor_role:
          type: string
          enum: [system, moderator, ops, admin]
        action_type:
          type: string
          enum: [nudge, block, quarantine, thread_lock, cooldown, restriction, override, appeal_grant]
        target_user_id:
          type: string
        target_scope:
          type: string
          enum: [message, thread, account]
        reason_code:
          type: string
        metadata:
          type: object
        is_permanent:
          type: boolean
          default: false

    ModerationAction:
      type: object
      required: [id, case_id, actor_id, actor_role, action_type, target_user_id, target_scope, reason_code, created_at]
      properties:
        id:
          type: string
          format: uuid
        case_id:
          type: string
          format: uuid
        actor_id:
          type: string
        actor_role:
          type: string
        action_type:
          type: string
        target_user_id:
          type: string
        target_scope:
          type: string
        reason_code:
          type: string
        metadata:
          type: object
        is_permanent:
          type: boolean
        requires_human:
          type: boolean
        approved_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    AppealRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          maxLength: 2000

    QueueStats:
      type: object
      properties:
        open_cases:
          type: integer
        in_review_cases:
          type: integer
        resolved_today:
          type: integer
        avg_resolution_hours:
          type: number
        false_positive_rate:
          type: number
        top_labels:
          type: array
          items:
            type: object
            properties:
              label:
                type: string
              count:
                type: integer